pyproject.toml
[project]
name = "sge-project"
version = "0.1.0"
description = ""
authors = [
    {name = "MateusLinhares-Dev",email = "mateuslinhares16@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.1.4,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "langchain (>=0.3.18,<0.4.0)",
    "openai (>=1.63.1,<2.0.0)",
    "langchain-openai (>=0.3.6,<0.4.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "huggingface-hub (>=0.28.1,<0.29.0)"
]

[tool.poetry.dependencies]
django = "^5.1.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
isort = "^5.13.2"
mkdocs = "^1.6.1"
prospector = "^1.13.3"
black = "^24.10.0"
flake8 = "6.0.0"
mypy = "^1.14.1"
pylint-django = "^2.6.1"
pip-audit = "^2.7.3"

[tool.poetry.scripts]
manage-django = 'sge_project.scripts.execute:run_manage'


